Various fixes and improvements.
Most notably, make sure the algorithm doesn't get stuck in an endless
loop. This can for instance happen if a branch is attracted by two nodes
and crosses the imaginary line connecting them. If in the next round the
tip of the branch is attracted by the same sources, the branch will grow
backwards again. To fix this, we remove branchlets whose tips overlaps
with already existing vein nodes.